pokemon_entities/requirements.txt
jango==3.1.14
folium==0.11.0
